+++ lib/floki/selector.ex
@@ -6,22 +6,23 @@ defmodule Floki.Selector do
   alias Floki.Selector
   alias Floki.AttributeSelector
 
-  defstruct id: nil, type: nil, classes: [], attributes: [], combinator: nil
+  defstruct id: nil, type: nil, classes: [], attributes: [], combinator: nil, namespace: nil
 
   @doc """
   Returns if a given node matches with a given selector.
   """
-  def match?(_node, %Selector{id: nil, type: nil, classes: [], attributes: []}) do
+  def match?(_node, %Selector{id: nil, type: nil, classes: [], attributes: [], namespace: nil}) do
     false
   end
   def match?(nil, _selector), do: false
   def match?({:comment, _comment}, _selector), do: false
   def match?({:pi, _xml, _xml_attrs}, _selector), do: false
-  def match?(node, selector) do
-    id_match?(node, selector.id)
-      && type_match?(node, selector.type)
-      && classes_matches?(node, selector.classes)
-      && attributes_matches?(node, selector.attributes)
+  def match?(html_node, selector) do
+    id_match?(html_node, selector.id)
+      && namespace_match?(html_node, selector.namespace)
+      && type_match?(html_node, selector.type)
+      && classes_matches?(html_node, selector.classes)
+      && attributes_matches?(html_node, selector.attributes)
   end
 
   defp id_match?(_node, nil), do: true
@@ -35,9 +36,26 @@ defmodule Floki.Selector do
     end
   end
 
+  defp namespace_match?(_node, nil), do: true
+  defp namespace_match?(_node, "*"), do: true
+  defp namespace_match?({type_maybe_with_namespace, _, _}, namespace) do
+    case String.split(type_maybe_with_namespace, ":") do
+      [ns, _type] ->
+        ns == namespace
+      [_type] -> false
+    end
+  end
+
   defp type_match?(_node, nil), do: true
-  defp type_match?({type, _, _}, type), do: true
-  defp type_match?({_type, _, _}, "*"), do: true
+  defp type_match?(_node, "*"), do: true
+  defp type_match?({type_maybe_with_namespace, _, _}, type) do
+    case String.split(type_maybe_with_namespace, ":") do
+      [_ns, tp] ->
+        tp == type
+      [tp] ->
+        tp == type
+    end
+  end
   defp type_match?(_, _), do: false
 
   defp classes_matches?(_node, []), do: true