+++ CHANGELOG.md
@@ -5,6 +5,69 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 
+## [0.10.0] - 2016-08-05
+
+### Changed
+
+- Change the search for namespaced elements using the correct CSS3 syntax.
+
+### Fixed
+
+- Fix the search for child elements when is more than two elements deep - thanks @gmile
+
+## [0.9.0] - 2016-06-16
+
+### Added
+
+- A separator between text when getting text from nodes - thanks @rochdi.
+
+## [0.8.1] - 2016-05-20
+
+### Added
+
+- Support rendering boolean attributes on `Floki.raw_html/1` - thanks @iamvery.
+
+### Changed
+
+- Update Mochiweb HTML parser dependency to version 2.15.0.
+
+## [0.8.0] - 2016-03-06
+
+### Added
+
+- Add possibility to search tags with namespaces.
+- Accept `Floki.Selector` as parameter of `Floki.find/2` instead of only strings - thanks @hansihe.
+
+### Changed
+
+- Using a smaller package with only the mochiweb HTML parser.
+
+## [0.7.2] - 2016-02-23
+
+### Fixed
+
+- Replace `<br>` nodes by newline (`\n`) in `DeepText` - thanks @maxneuvians.
+- Allow `FilterOut` to filter special nodes, like `comment`.
+
+## [0.7.1] - 2015-11-14
+
+### Fixed
+
+- Ignore PHP scripts when finding nodes.
+
+## [0.7.0] - 2015-11-03
+
+### Added
+
+- Add support for excluding script notes in `Floki.text`.
+By default, it will exclude those nodes, but it can be enabled with
+the flag `js: true` - thanks @vikeri!
+
+### Fixed
+
+- Fix find for sibling nodes when the precendent selector match an element
+at the end of sibling list - fix [issue #39](https://github.com/philss/floki/issues/39)
+
 ## [0.6.1] - 2015-10-11
 
 ### Fixed
@@ -133,7 +196,14 @@ of the parent element inside HTML.
 
 - Elixir version requirement from "~> 1.0.0" to ">= 1.0.0".
 
-[unreleased]: https://github.com/philss/floki/compare/v0.6.1...HEAD
+[unreleased]: https://github.com/philss/floki/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/philss/floki/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/philss/floki/compare/v0.8.1...v0.9.0
+[0.8.1]: https://github.com/philss/floki/compare/v0.8.0...v0.8.1
+[0.8.0]: https://github.com/philss/floki/compare/v0.7.2...v0.8.0
+[0.7.2]: https://github.com/philss/floki/compare/v0.7.1...v0.7.2
+[0.7.1]: https://github.com/philss/floki/compare/v0.7.0...v0.7.1
+[0.7.0]: https://github.com/philss/floki/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/philss/floki/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/philss/floki/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/philss/floki/compare/v0.4.1...v0.5.0